+++ src/cmd/ship.go
@@ -1,6 +1,8 @@
 package cmd
 
 import (
+	"strings"
+
 	"github.com/Originate/exit"
 	"github.com/Originate/git-town/src/drivers"
 	"github.com/Originate/git-town/src/git"
@@ -12,20 +14,19 @@ import (
 	"github.com/spf13/cobra"
 )
 
-type shipAllConfig struct {
-	BranchToMergeInto string
-	BranchesToShip    []string
-	InitialBranch     string
+type shipConfig struct {
+	BranchToShip  string
+	InitialBranch string
 }
 
-var shipAllCommitMessage string
+var commitMessage string
 
-var shipAllCmd = &cobra.Command{
-	Use:   "shipstack",
-	Short: "Deliver a stack of feature branches",
-	Long: `Deliver a stack of feature branches
+var shipCmd = &cobra.Command{
+	Use:   "ship",
+	Short: "Deliver a completed feature branch",
+	Long: `Deliver a completed feature branch
 
-Squash-merges the stack of feature branches up to the current branch (or <branch_name> if given)
+Squash-merges the current branch, or <branch_name> if given,
 into the main branch, resulting in linear history on the main branch.
 
 - syncs the main branch
@@ -35,7 +36,9 @@ into the main branch, resulting in linear history on the main branch.
   with commit message specified by the user
 - pushes the main branch to the remote repository
 - deletes <branch_name> from the local and remote repositories
-- repeat until the stack has been merged
+
+Only shipping of direct children of the main branch is allowed.
+To ship a nested child branch, all ancestor branches have to be shipped or killed.
 
 If you are using GitHub, this command can squash merge pull requests via the GitHub API. Setup:
 1. Get a GitHub personal access token with the "repo" scope
@@ -43,8 +46,8 @@ If you are using GitHub, this command can squash merge pull requests via the Git
 Now anytime you ship a branch with a pull request on GitHub, it will squash merge via the GitHub API.
 It will also update the base branch for any pull requests against that branch.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		config := gitShipAllConfig(args)
-		stepList := getShipAllStepList(config)
+		config := gitShipConfig(args)
+		stepList := getShipStepList(config)
 		runState := steps.NewRunState("ship", stepList)
 		steps.Run(runState)
 	},
@@ -57,62 +60,69 @@ It will also update the base branch for any pull requests against that branch.`,
 	},
 }
 
-func gitShipAllConfig(args []string) (result shipAllConfig) {
+func gitShipConfig(args []string) (result shipConfig) {
 	result.InitialBranch = git.GetCurrentBranchName()
-	var branchToShip string
 	if len(args) == 0 {
-		branchToShip = result.InitialBranch
+		result.BranchToShip = result.InitialBranch
 	} else {
-		branchToShip = args[0]
+		result.BranchToShip = args[0]
 	}
-	if branchToShip == result.InitialBranch {
+	if result.BranchToShip == result.InitialBranch {
 		git.EnsureDoesNotHaveOpenChanges("Did you mean to commit them before shipping?")
 	}
 	if git.HasRemote("origin") && !git.IsOffline() {
 		script.Fetch()
 	}
-	if branchToShip != result.InitialBranch {
-		git.EnsureHasBranch(branchToShip)
+	if result.BranchToShip != result.InitialBranch {
+		git.EnsureHasBranch(result.BranchToShip)
 	}
-	git.EnsureIsFeatureBranch(branchToShip, "Only feature branches can be shipped.")
-	prompt.EnsureKnowsParentBranches([]string{branchToShip})
-	ancestors := git.GetAncestorBranches(branchToShip)
-	result.BranchToMergeInto = ancestors[0]
-	result.BranchesToShip = ancestors[1:]
+	git.EnsureIsFeatureBranch(result.BranchToShip, "Only feature branches can be shipped.")
+	prompt.EnsureKnowsParentBranches([]string{result.BranchToShip})
+	ensureParentBranchIsMainOrPerennialBranch(result.BranchToShip)
 	return
 }
 
-func getShipAllStepList(config shipAllConfig) steps.StepList {
+func ensureParentBranchIsMainOrPerennialBranch(branchName string) {
+	parentBranch := git.GetParentBranch(branchName)
+	if !git.IsMainBranch(parentBranch) && !git.IsPerennialBranch(parentBranch) {
+		ancestors := git.GetAncestorBranches(branchName)
+		ancestorsWithoutMainOrPerennial := ancestors[1:]
+		oldestAncestor := ancestorsWithoutMainOrPerennial[0]
+		util.ExitWithErrorMessage(
+			"Shipping this branch would ship "+strings.Join(ancestorsWithoutMainOrPerennial, ", ")+" as well.",
+			"Please ship \""+oldestAncestor+"\" first.",
+		)
+	}
+}
+
+func getShipStepList(config shipConfig) steps.StepList {
 	result := steps.StepList{}
 	var isOffline = git.IsOffline()
-	isShippingInitialBranch := config.BranchesToShip[len(config.BranchesToShip)-1] == config.InitialBranch
-	branchToMergeInto := config.BranchToMergeInto
-	for _, branchToShip := range config.BranchesToShip {
-		result.AppendList(steps.GetSyncBranchSteps(branchToMergeInto, true))
-		result.AppendList(steps.GetSyncBranchSteps(branchToShip, false))
-		result.Append(&steps.EnsureHasShippableChangesStep{BranchName: branchToShip})
-		result.Append(&steps.CheckoutBranchStep{BranchName: branchToMergeInto})
-		canShipWithDriver, defaultCommitMessage := getCanShipAllWithDriver(branchToShip, branchToMergeInto)
-		if canShipWithDriver {
-			result.Append(&steps.PushBranchStep{BranchName: branchToShip})
-			result.Append(&steps.DriverMergePullRequestStep{BranchName: branchToShip, CommitMessage: shipAllCommitMessage, DefaultCommitMessage: defaultCommitMessage})
-			result.Append(&steps.PullBranchStep{})
-		} else {
-			result.Append(&steps.SquashMergeBranchStep{BranchName: branchToShip, CommitMessage: shipAllCommitMessage})
-		}
-		if git.HasRemote("origin") && !isOffline {
-			result.Append(&steps.PushBranchStep{BranchName: branchToMergeInto, Undoable: true})
-		}
-		childBranches := git.GetChildBranches(branchToShip)
-		if canShipWithDriver || (git.HasTrackingBranch(branchToShip) && len(childBranches) == 0 && !isOffline) {
-			// I don't think I need this step. The Github API seems to be deleting remote branches.
-			// result.Append(&steps.DeleteRemoteBranchStep{BranchName: branchToShip, IsTracking: true})
-		}
-		result.Append(&steps.DeleteLocalBranchStep{BranchName: branchToShip})
-		result.Append(&steps.DeleteParentBranchStep{BranchName: branchToShip})
-		for _, child := range childBranches {
-			result.Append(&steps.SetParentBranchStep{BranchName: child, ParentBranchName: branchToMergeInto})
-		}
+	branchToMergeInto := git.GetParentBranch(config.BranchToShip)
+	isShippingInitialBranch := config.BranchToShip == config.InitialBranch
+	result.AppendList(steps.GetSyncBranchSteps(branchToMergeInto, true))
+	result.AppendList(steps.GetSyncBranchSteps(config.BranchToShip, false))
+	result.Append(&steps.EnsureHasShippableChangesStep{BranchName: config.BranchToShip})
+	result.Append(&steps.CheckoutBranchStep{BranchName: branchToMergeInto})
+	canShipWithDriver, defaultCommitMessage := getCanShipWithDriver(config.BranchToShip, branchToMergeInto)
+	if canShipWithDriver {
+		result.Append(&steps.PushBranchStep{BranchName: config.BranchToShip})
+		result.Append(&steps.DriverMergePullRequestStep{BranchName: config.BranchToShip, CommitMessage: commitMessage, DefaultCommitMessage: defaultCommitMessage})
+		result.Append(&steps.PullBranchStep{})
+	} else {
+		result.Append(&steps.SquashMergeBranchStep{BranchName: config.BranchToShip, CommitMessage: commitMessage})
+	}
+	if git.HasRemote("origin") && !isOffline {
+		result.Append(&steps.PushBranchStep{BranchName: branchToMergeInto, Undoable: true})
+	}
+	childBranches := git.GetChildBranches(config.BranchToShip)
+	if canShipWithDriver || (git.HasTrackingBranch(config.BranchToShip) && len(childBranches) == 0 && !isOffline) {
+		result.Append(&steps.DeleteRemoteBranchStep{BranchName: config.BranchToShip, IsTracking: true})
+	}
+	result.Append(&steps.DeleteLocalBranchStep{BranchName: config.BranchToShip})
+	result.Append(&steps.DeleteParentBranchStep{BranchName: config.BranchToShip})
+	for _, child := range childBranches {
+		result.Append(&steps.SetParentBranchStep{BranchName: child, ParentBranchName: branchToMergeInto})
 	}
 	if !isShippingInitialBranch {
 		result.Append(&steps.CheckoutBranchStep{BranchName: config.InitialBranch})
@@ -121,7 +131,7 @@ func getShipAllStepList(config shipAllConfig) steps.StepList {
 	return result
 }
 
-func getCanShipAllWithDriver(branch, parentBranch string) (bool, string) {
+func getCanShipWithDriver(branch, parentBranch string) (bool, string) {
 	if !git.HasRemote("origin") {
 		return false, ""
 	}
@@ -138,6 +148,6 @@ func getCanShipAllWithDriver(branch, parentBranch string) (bool, string) {
 }
 
 func init() {
-	shipAllCmd.Flags().StringVarP(&shipAllCommitMessage, "message", "m", "", "Specify the commit message for the squash commit")
-	RootCmd.AddCommand(shipAllCmd)
+	shipCmd.Flags().StringVarP(&commitMessage, "message", "m", "", "Specify the commit message for the squash commit")
+	RootCmd.AddCommand(shipCmd)
 }
